Added save path and save skip
The output path for the resulting grid data can be changed by
adding "path=XY" as commandline argument. When it is not provided
the path defaults to "./results".
By setting config.gridConfig.skip_save the saving of datapoints can be
skipped. e.g if set to 100 only every 100th datapoint will be saved